+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v3.0.0
+* allow to scroll elements
+
+# v2.1.1
+* improve invalid position handling
+
 # v2.1.0
 * now aborts early if a scroll position is outside of the document
 * add ability to abort programmatic scrolling whenever a user scrolls